displays character on the map moving in the correct direction
