+++ src/components/map/Map.js
@@ -2,17 +2,18 @@ import React from 'react';
 import { SPRITE_SIZE } from '../../config/constants';
 import { connect } from 'react-redux';
 
-import 'styles.css';
+
+import './styles.css';
 
 function getTileSprite(type) {
     switch(type) {
         case 0:
-            return 'grass'
-        case 0:
+            return 'tile'
+        case 1:
             return 'tree'
-        case 0:
+        case 2:
             return 'rock'
-        case 0:
+        case 3:
             return 'chest'
     }
 }
@@ -42,16 +43,16 @@ function MapRow(props) {
     </div>
 }
 
-export default function Map(props) {
-    return(
-        <div
-        style={{
+function Map(props) {
+    return (
+        <div 
+            style={{
             position: 'relative',
             top: '0px',
             left: '0px',
-            width: '800px',
-            height: '400px',
-            border: '4px solid white'
+            width: '640px',
+            height: '320px',
+            border: '4px solid white',
         }}
         >
         {
@@ -61,3 +62,11 @@ export default function Map(props) {
     )
 }
 
+// export default Map;
+function mapStateToProps(state) {
+    return {
+        tiles: state.map.tiles
+    }
+}
+
+export default connect(mapStateToProps)(Map);
