+++ src/components/player/PlayerMovement.js
@@ -7,37 +7,41 @@ export default function PlayerMovement(player) {
         switch(direction) {
             case 'WEST':
                 return [ oldPos[0]-SPRITE_SIZE, oldPos[1] ]
+
             case 'EAST':
                 return [ oldPos[0]+SPRITE_SIZE, oldPos[1] ]
+
             case 'NORTH':
                 return [  oldPos[0], oldPos[1]-SPRITE_SIZE ]
+
             case 'SOUTH':
                 return [  oldPos[0], oldPos[1]+SPRITE_SIZE ]
+
         }
-    };
+    }
 
     function getSpriteLocation(direction, walkIndex) {
         switch(direction) {
-            case 'WEST':
-                return `${SPRITE_SIZE*walkIndex}px ${SPRITE_SIZE*2}px`
-            case 'EAST':
-                return `${SPRITE_SIZE*walkIndex}px ${SPRITE_SIZE*1}px`
-            case 'NORTH':
-                return `${SPRITE_SIZE*walkIndex}px ${SPRITE_SIZE*3}px`
             case 'SOUTH':
                 return `${SPRITE_SIZE*walkIndex}px ${SPRITE_SIZE*0}px`
+            case 'NORTH':
+                return `${SPRITE_SIZE*walkIndex}px ${SPRITE_SIZE*1}px`
+            case 'EAST':
+                return `${SPRITE_SIZE*walkIndex}px ${SPRITE_SIZE*2}px`
+            case 'WEST':
+                return `${SPRITE_SIZE*walkIndex}px ${SPRITE_SIZE*3}px`
         }
-    };
+    }
 
     function getWalkIndex() {
         const walkIndex = store.getState().player.walkIndex
-        return walkIndex >= 7 ? 0 : walkIndex + 1
-    };
+        return walkIndex >= 2 ? 0 : walkIndex + 1
+    }
 
     function observeBoundaries(oldPos, newPos) {
         return (newPos[0] >= 0 && newPos[0] <= MAP_WIDTH - SPRITE_SIZE) &&
                 (newPos[1] >= 0 && newPos[1] <= MAP_HEIGHT - SPRITE_SIZE)
-    };
+    }
 
     function observeImpassable(oldPos, newPos) {
         const tiles = store.getState().map.tiles
@@ -45,7 +49,7 @@ export default function PlayerMovement(player) {
         const x = newPos[0] / SPRITE_SIZE
         const nextTile = tiles[y][x]
         return nextTile < 5
-    };
+    }
 
     function dispatchMove(direction, newPos) {
         const walkIndex = getWalkIndex()
@@ -58,15 +62,14 @@ export default function PlayerMovement(player) {
                 spriteLocation: getSpriteLocation(direction, walkIndex),
             }
         });
-    };
+    }
 
     function attemptMove(direction) {
         const oldPos = store.getState().player.position;
-        const newPos = getNewPosition(oldPos, newPos);
-
-        if(observeBoundaries(oldPos, newPos) && observeImpassable(oldPos, newPos))
+        const newPos = getNewPosition(oldPos, direction);
+        if (observeBoundaries(oldPos, newPos) && observeImpassable(oldPos, newPos))
             dispatchMove(direction, newPos)
-    };
+    }
 
     function handleKeyDown(e) {
         e.preventDefault()
@@ -83,11 +86,11 @@ export default function PlayerMovement(player) {
             default:
                 console.log(e.keyCode)
         }
-    };
+    }
 
     window.addEventListener('keydown', (e) => {
         handleKeyDown(e)
-    });
+    })
 
-    return (player)
+    return player
 }
