+++ src/config/constants.js
@@ -1,3 +1,26 @@
 export const SPRITE_SIZE=32;
 export const MAP_HEIGHT = SPRITE_SIZE * 10
 export const MAP_WIDTH = SPRITE_SIZE * 20
+
+export const tiles = [
+    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    // [0, 6, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 6, 5, 5 ,5],
+    // [0, 6, 6, 6, 0, 0, 0, 5, 0, 0, 0, 5, 0, 0, 6, 3, 3, 3, 5 ,6],
+    // [0, 0, 6, 0, 5, 0, 0, 0, 0, 6, 5, 5, 0, 0, 6, 3, 6, 3, 5 ,6],
+    // [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 5, 0, 0, 0, 3, 6, 3, 6 ,4],
+    // [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 5, 5, 0, 0, 0, 6, 0, 0 ,0],
+    // [0, 0, 0, 0, 0, 0, 0, 5, 0, 0, 6, 6, 6, 6, 6, 0, 6, 6, 6 ,6],
+    // [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 6, 0, 0, 0, 0, 0 ,0],
+    // [0, 6, 6, 6, 5, 0, 5, 0, 0, 0, 0, 6, 0, 0, 0, 0, 0, 0, 0 ,0],
+    // [6, 6, 6, 6, 6, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0],
+    // [6, 6, 6, 6, 6, 6, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 ,0]
+]
