+++ src/containers/MapContainer.js
@@ -1,10 +1,10 @@
-import { connect } from 'react-redux';
-import Map from '../components/map/Map';
+// import { connect } from 'react-redux';
+// import Map from '../components/map/Map';
 
-function mapStateToProps(state) {
-    return {
-        tiles: state.map.tiles
-    }
-}
+// function mapStateToProps(state) {
+//     return {
+//         tiles: state.map.tiles
+//     }
+// }
 
-export default connect(mapStateToProps)(Map);
+// export default connect(mapStateToProps)(Map);
